add assassin and excavator

+++ Assets/Scripts/Main/Card.cs
@@ -14,7 +14,7 @@ public class Card : PhotonCompatible
 
     public string extraText { get; protected set; }
     public int coinCost { get; protected set; }
-    protected int abilityValue;
+    protected float abilityValue;
 
     #endregion
